+++ test/internal/testService.js
@@ -1,4 +1,4 @@
-const { requestContext } = require('../../index')
+const { requestContext } = require('../..')
 
 // Test class to check if nested calls with promises work correctly with async local storage
 class TestService {
